requirements.txt
requests==2.25.0
python-telegram-bot==13.1
dnspython==2.0.0